Unify gutterOption values and remove duplicate code
