+++ src/blocks/gallery-offset/inspector.js
@@ -5,6 +5,7 @@ import GalleryLinkSettings from '../../components/block-gallery/gallery-link-set
 import captionOptions from '../../components/block-gallery/options/caption-options';
 import SizeControl from '../../components/size-control';
 import OptionSelectorControl from '../../components/option-selector-control';
+import gutterOptions from '../../utils/gutter-options';
 
 /**
  * WordPress dependencies
@@ -74,38 +75,6 @@ class Inspector extends Component {
 			captions,
 			captionStyle,
 		} = attributes;
-
-		const gutterOptions = [
-			{
-				value: 'no',
-				label: __( 'None', 'coblocks' ),
-				shortName: __( 'None', 'coblocks' ),
-			},
-			{
-				value: 'small',
-				/* translators: abbreviation for small size */
-				label: __( 'S', 'coblocks' ),
-				tooltip: __( 'Small', 'coblocks' ),
-			},
-			{
-				value: 'medium',
-				/* translators: abbreviation for medium size */
-				label: __( 'M', 'coblocks' ),
-				tooltip: __( 'Medium', 'coblocks' ),
-			},
-			{
-				value: 'large',
-				/* translators: abbreviation for large size */
-				label: __( 'L', 'coblocks' ),
-				tooltip: __( 'Large', 'coblocks' ),
-			},
-			{
-				value: 'huge',
-				/* translators: abbreviation for largest size */
-				label: __( 'XL', 'coblocks' ),
-				tooltip: __( 'Huge', 'coblocks' ),
-			},
-		];
 
 		return (
 			<InspectorControls>
