+++ src/blocks/row/inspector.js
@@ -12,6 +12,7 @@ import applyWithColors from './colors';
 import { BackgroundPanel } from '../../components/background';
 import DimensionsControl from '../../components/dimensions-control';
 import OptionSelectorControl from '../../components/option-selector-control';
+import gutterOptions from '../../utils/gutter-options';
 
 /**
  * WordPress dependencies
@@ -97,33 +98,6 @@ class Inspector extends Component {
 			paddingUnit,
 			hasMarginControl,
 		} = attributes;
-
-		const gutterOptions = [
-			{
-				value: 'small',
-				/* translators: abbreviation for small size */
-				label: __( 'S', 'coblocks' ),
-				tooltip: __( 'Small', 'coblocks' ),
-			},
-			{
-				value: 'medium',
-				/* translators: abbreviation for medium size */
-				label: __( 'M', 'coblocks' ),
-				tooltip: __( 'Medium', 'coblocks' ),
-			},
-			{
-				value: 'large',
-				/* translators: abbreviation for large size */
-				label: __( 'L', 'coblocks' ),
-				tooltip: __( 'Large', 'coblocks' ),
-			},
-			{
-				value: 'huge',
-				/* translators: abbreviation for largest size */
-				label: __( 'XL', 'coblocks' ),
-				tooltip: __( 'Huge', 'coblocks' ),
-			},
-		];
 
 		let selectedRows = 1;
