+++ src/blocks/services/inspector.js
@@ -8,6 +8,7 @@ import classnames from 'classnames';
  */
 import HeadingToolbar from '../../components/heading-toolbar';
 import OptionSelectorControl from '../../components/option-selector-control';
+import gutterOptions from '../../utils/gutter-options';
 
 /**
  * WordPress dependencies.
@@ -27,33 +28,6 @@ const Inspector = ( props ) => {
 		onToggleCtas,
 		onUpdateStyle,
 	} = props;
-
-	const gutterOptions = [
-		{
-			value: 'small',
-			/* translators: abbreviation for small size */
-			label: __( 'S', 'coblocks' ),
-			tooltip: __( 'Small', 'coblocks' ),
-		},
-		{
-			value: 'medium',
-			/* translators: abbreviation for medium size */
-			label: __( 'M', 'coblocks' ),
-			tooltip: __( 'Medium', 'coblocks' ),
-		},
-		{
-			value: 'large',
-			/* translators: abbreviation for large size */
-			label: __( 'L', 'coblocks' ),
-			tooltip: __( 'Large', 'coblocks' ),
-		},
-		{
-			value: 'huge',
-			/* translators: abbreviation for largest size */
-			label: __( 'XL', 'coblocks' ),
-			tooltip: __( 'Huge', 'coblocks' ),
-		},
-	];
 
 	return (
 		<InspectorControls>
